Modificacion de ids (int -> Long). Creacion de secciones predeterminadas al registrar un usuario.

+++ src/main/java/com/burgosmanuel/portfolio/educacion/EducacionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EducacionRepository extends JpaRepository<Educacion, Integer>{
+public interface EducacionRepository extends JpaRepository<Educacion, Long>{
     
 }
