+++ src/main/java/com/burgosmanuel/portfolio/habilidad/HabilidadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HabilidadRepository extends JpaRepository<Habilidad, Integer>{
+public interface HabilidadRepository extends JpaRepository<Habilidad, Long>{
     
 }
